% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asCAGEfightR.R
\name{asCAGEfightR}
\alias{asCAGEfightR}
\title{Transforms a CAGEexp object into a CAGEfightR-friendly object}
\usage{
asCAGEfightR(ce)
}
\arguments{
\item{ce}{a CAGEexp object}
}
\value{
A \code{RangedSummarizedExperiment} object
}
\description{
Transforms a CAGEexp object into a \code{RangedSummarizedExperiment} object, which can be used with various \code{CAGEfightR} functions.
The output is similar to that of \code{\link[CAGEfightR:quantifyCTSSs]{CAGEfightR::quantifyCTSSs()}}.
}
\examples{
\dontrun{
asCAGEfightR(ce)
}

}
