% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeBed.R
\name{makeBed}
\alias{makeBed}
\title{Exports clusters from CAGEexp object}
\usage{
makeBed(ce, sl_found)
}
\arguments{
\item{ce}{CAGEexp object}

\item{sl_found}{boolean, whether splice leader has been found in data}
}
\value{
tracks in bed format
}
\description{
Exports clusters from CAGEexp objects containing splice leader or splice leader removed data to black and white bed tracks.
For clusters, call \link[CAGErAid]{quickClustersSL} or \link[CAGErAid]{quickClustersNO}.
}
\examples{
\dontrun{
makeBed(ce, sl_found = TRUE)
}
}
