% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqcGenstats.R
\name{mqcGenstats}
\alias{mqcGenstats}
\title{Imports general QC data (multiqc_general_stats.txt) and writes it into the CAGEexp object
Should be ran after quickStats and mqcHisat2}
\usage{
mqcGenstats(ce, qc)
}
\arguments{
\item{ce}{CAGEexp object}

\item{qc}{table with multiqc_general_stats.txt data}
}
\value{
CAGEexp object
}
\description{
Imports general QC data (multiqc_general_stats.txt) and writes it into the CAGEexp object
Should be ran after quickStats and mqcHisat2
}
\examples{
\dontrun{
mqcGenstats(ce, qc)
}

}
