% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mqcHisat2.R
\name{mqcHisat2}
\alias{mqcHisat2}
\title{Imports Hisat2 QC data (multiqc_hisat2.txt) and writes it into the CAGEexp object}
\usage{
mqcHisat2(ce, hisat, check_multimap = FALSE)
}
\arguments{
\item{ce}{CAGEexp object}

\item{hisat}{table with multiqc_hisat2.txt data}

\item{check_multimap}{optional; checks if all reads have been removed by the pipeline}
}
\value{
CAGEexp object
}
\description{
Imports Hisat2 QC data (multiqc_hisat2.txt) and writes it into the CAGEexp object
}
\examples{
\dontrun{
mqcHisat2(ce, hisat, check_multimap = TRUE)
}
}
