% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickCC.R
\name{quickCC}
\alias{quickCC}
\title{Aggregates tag clusters into consensus clusters}
\usage{
quickCC(ce, sl_found, gff)
}
\arguments{
\item{ce}{CAGEexp object}

\item{sl_found}{boolean, whether data contains samples with or without SL sequences}

\item{gff}{GFF file}
}
\value{
CAGEexp object
}
\description{
Quickly aggregates CTSSs into consensus clusters from data from samples with or without splice leader detected with established parameters.
For finding tag clusters, use \code{\link[=quickTC]{quickTC()}}.

The method for clustering CTSSs in samples without SL detected is distclu, and with - paraclu. By default, the algorithm uses 4 cores.

The function annotates found consensus clusters with the use of a GFF file. For easy processing of gff files use \code{\link[=quickGFF]{quickGFF()}}.
}
\examples{
\dontrun{
quickCC(ce, sl_found = TRUE, gff)
}
}
