% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickClustersNO.R
\name{quickClustersNO}
\alias{quickClustersNO}
\title{Clusters CTSSs for samples without splice leader}
\usage{
quickClustersNO(ce, aggregate = FALSE, gff = NULL)
}
\arguments{
\item{ce}{CAGEexp object}

\item{aggregate}{whether to return consensus clusters}

\item{gff}{GFF file}
}
\value{
CAGEexp object
}
\description{
Quickly clusters CTSSs in data from samples without splice leader detected with established parameters.
For clustering CTSSs in samples with splice leader found, use \code{\link[=quickClustersSL]{quickClustersSL()}}.

The method for clustering CTSSs is distclu, and by default the algorithm uses 4 cores.

Optional aggregating tags is possible with aggregate = TRUE, at which the user needs to provide a GFF file.
For easy processing of gff files use \code{\link[=quickGFF]{quickGFF()}}.
}
\examples{
\dontrun{
quickClustersNO(ce, aggregate = TRUE, gff)
}
}
