% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickClustersSL.R
\name{quickClustersSL}
\alias{quickClustersSL}
\title{Quickly clusters CTSSs in splice leader-containing sample data with established parameters.
For clustering CTSSs in samples without splice leader found, use \link[CAGErAid]{quickClustersNO}.}
\usage{
quickClustersSL(ce, aggregate = FALSE, gff = NULL)
}
\arguments{
\item{ce}{CAGEexp object}

\item{aggregate}{whether to return consensus clusters}

\item{gff}{GFF file}
}
\value{
CAGEexp object
}
\description{
The method for clustering CTSSs is paraclu, and by default the algorithm uses 4 cores.
}
\details{
Optional aggregating tags is possible with aggregate = TRUE, at which the user needs to provide a GFF file.
For easy processing of gff files use \link[CAGErAid]{quickGFF}.
}
\examples{
\dontrun{
quickClustersSL(ce, aggregate = TRUE, gff)
}
}
