% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickGFF.R
\name{quickGFF}
\alias{quickGFF}
\title{Quickly imports and processes a GFF object as a character vector for use in annotating CAGEexp objects}
\usage{
quickGFF(gff_path, types = c("transcript", "intron", "exon"))
}
\arguments{
\item{gff_path}{path to a GFF file}

\item{types}{types of features of interest}
}
\value{
GRanges object with selected features
}
\description{
Quickly imports and processes a GFF object as a character vector for use in annotating CAGEexp objects
}
\examples{
\dontrun{
quickGFF('barcelona.gff', types = c('intron', 'exon'))
}
}
