% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPolish.R
\name{quickPolish}
\alias{quickPolish}
\title{Polish CAGEexp data}
\usage{
quickPolish(ce)
}
\arguments{
\item{ce}{CAGEexp object}
}
\value{
CAGEexp object
}
\description{
Fixes sample names and adds metadata. Use before \code{\link[=quickMQC]{quickMQC()}}.
}
\examples{
\dontrun{
quickPolish(ce)
}
}
