% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickTC.R
\name{quickTC}
\alias{quickTC}
\title{Calculates tag clusters with set parameters}
\usage{
quickTC(ce, sl_found)
}
\arguments{
\item{ce}{CAGEexp object}

\item{sl_found}{boolean, whether data contains samples with or without SL sequences}
}
\value{
CAGEexp object
}
\description{
Quickly calculates tag clusters from CTSSs in data from samples with or without splice leader detected with established parameters.
For finding consensus clusters, use \code{\link[=quickCC]{quickCC()}}.

The method for clustering CTSSs in samples without SL detected is distclu, and with - paraclu. By default, the algorithm uses 4 cores.
}
\examples{
\dontrun{
quickTC(ce, sl_found = TRUE)
}
}
