% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickClustersNO.R
\name{quickClustersNO}
\alias{quickClustersNO}
\title{Quickly clusters CTSSs in data from samples without splice leader detected with established parameters.
For clustering CTSSs in samples with splice leader found, use \link[CAGErAid]{quickClustersSL}.}
\usage{
quickClustersNO(ce, aggregate = FALSE, gff = NULL)
}
\arguments{
\item{ce}{CAGEexp object}

\item{aggregate}{whether to return consensus clusters}

\item{gff}{GFF file}
}
\value{
CAGEexp object
}
\description{
The method for clustering CTSSs is distclu, and by default the algorithm uses 4 cores.
}
\details{
Optional aggregating tags is possible with aggregate = TRUE, at which the user needs to provide a GFF file.
For easy processing of gff files use \link[CAGErAid]{quickGFF}.
}
\examples{
\dontrun{
quickClustersNO(ce, aggregate = TRUE, gff)
}
}
