% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMQC.R
\name{quickMQC}
\alias{quickMQC}
\title{Imports QC data}
\usage{
quickMQC(ce, guess_path = TRUE, hisat = NULL, qc = NULL)
}
\arguments{
\item{ce}{CAGEexp object}

\item{guess_path}{boolean, if TRUE, checks if the multiqc_hisat2.txt and multiqc_general_stats.txt paths can be inferred from the input files in the CAGEexp object}

\item{hisat}{table with multiqc_hisat2.txt data}

\item{qc}{table with multiqc_general_stats.txt data}
}
\value{
CAGEexp object
}
\description{
Imports Hisat2 QC (multiqc_hisat2.txt) and general QC data (multiqc_general_stats.txt) and writes it into the CAGEexp object
}
\examples{
\dontrun{
quickMQC(ce, guess_path = FALSE, hisat, qc)
}

}
