% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickMQC.R
\name{quickMQC}
\alias{quickMQC}
\title{Imports Hisat2 QC (multiqc_hisat2.txt) and general QC data (multiqc_general_stats.txt) and writes it into the CAGEexp object}
\usage{
quickMQC(ce, hisat, qc, check_multimap = FALSE)
}
\arguments{
\item{ce}{CAGEexp object}

\item{hisat}{table with multiqc_hisat2.txt data}

\item{qc}{table with multiqc_general_stats.txt data}

\item{check_multimap}{optional; checks if all reads have been removed by the pipeline}
}
\value{
CAGEexp object
}
\description{
Imports Hisat2 QC (multiqc_hisat2.txt) and general QC data (multiqc_general_stats.txt) and writes it into the CAGEexp object
}
\examples{
\dontrun{
quickMQC(ce, hisat, qc, check_multimap = TRUE)
}
}
