% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickPolish.R
\name{quickPolish}
\alias{quickPolish}
\title{Polish CAGEexp data}
\usage{
quickPolish(ce, reads = c("bar", "oki", "osa"))
}
\arguments{
\item{ce}{CAGEexp object}

\item{reads}{Barcelona, Okinawa or Osaka reads}
}
\value{
CAGEexp object
}
\description{
Fixes sample names and adds metadata. Use before \code{\link[=quickMQC]{quickMQC()}}.
}
\examples{
\dontrun{
quickPolish(ce, reads = 'bar')
}
}
